django/requirements.txt
asgiref==3.2.7
astroid==2.3.3
certifi==2020.4.5.1
chardet==3.0.4
Django==2.2
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
ptvsd==4.3.2
pylint==2.4.4
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.11.2
